Add functionality to retrieve and store granule date ranges
